requirements.txt
streamlit
pandas
plotly
scipy
pymodbus
psycopg2-binary
scikit-learn
xgboost
